+++ jdkringbuffer-core/src/main/java/org/jdkstack/jdkringbuffer/core/AbstractBlockingQueue.java
@@ -5,7 +5,6 @@ import java.util.Collection;
 import java.util.Iterator;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.LongAdder;
 
 /**
  * 阻塞队列核心.
