+++ jdkringbuffer-core/src/main/java/org/jdkstack/jdkringbuffer/core/AbstractLockBlockingQueueV1.java
@@ -1,30 +1,26 @@
-package org.jdkstack.jdkringbuffer.core.spsc.version4;
+package org.jdkstack.jdkringbuffer.core;
 
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.locks.LockSupport;
 import org.jdkstack.jdkringbuffer.api.RingBufferBlockingQueue;
-import org.jdkstack.jdkringbuffer.core.AbstractBlockingQueue;
-import org.jdkstack.jdkringbuffer.core.Constants;
 
-/**
- * 单生产单消费SPSC阻塞队列.
- *
- * <p>不需要处理线程安全,不使用CAS锁.
- *
- * @author admin
- * @param <E> e .
- */
-public abstract class AbstractSpscBlockingQueueV2<E> extends AbstractBlockingQueue<E>
+public abstract class AbstractLockBlockingQueueV1<E> extends AbstractBlockingQueue<E>
     implements BlockingQueue<E>, RingBufferBlockingQueue {
   /** 环形数组. */
-  private final E[] ringBuffer;
+  protected final E[] ringBuffer;
+  /** 环形数组入队时,是否被其他线程抢先放入了值. */
+  protected final AtomicInteger tailLock = new AtomicInteger(0);
+  /** 环形数组出队时,是否被其他线程抢先获取了值. */
+  protected final AtomicInteger headLock = new AtomicInteger(0);
 
-  protected AbstractSpscBlockingQueueV2() {
+  protected AbstractLockBlockingQueueV1() {
     this(Constants.CAPACITY);
   }
 
   @SuppressWarnings("unchecked")
-  protected AbstractSpscBlockingQueueV2(final int capacity) {
+  protected AbstractLockBlockingQueueV1(final int capacity) {
     super(capacity, capacity - 1);
     // jdk 泛型数组,会有检查异常,但不影响什么,用unchecked关闭检查.
     this.ringBuffer = (E[]) new Object[capacity];
@@ -45,13 +41,20 @@ public abstract class AbstractSpscBlockingQueueV2<E> extends AbstractBlockingQue
     final int tailSeq = this.tail.get();
     final int queueStart = tailSeq - this.capacity;
     boolean flag = false;
-    // 检查环形数组是否满了.
+    // 检查环形数组是否满了. 检查是否被其他线程修改.
     if (0 > queueStart || this.head.get() > queueStart) {
-      // 向环形数组设置元素,取模后向对应的下标设置元素.
-      final int tailSlot = tailSeq & index;
-      this.ringBuffer[tailSlot] = e;
-      this.tail.getAndIncrement();
-      flag = true;
+      if (this.tailLock.compareAndSet(tailSeq, tailSeq + 1)) {
+        // 向环形数组设置元素,取模后向对应的下标设置元素.
+        final int tailSlot = tailSeq & index;
+        this.ringBuffer[tailSlot] = e;
+        this.tail.getAndIncrement();
+        flag = true;
+      } else {
+        // 如果被占用,暂停1nanos.
+        // https://blogs.oracle.com/dave/lightweight-contention-
+        // management-for-efficient-compare-and-swap-operations
+        LockSupport.parkNanos(1L);
+      }
     }
     return flag;
   }
@@ -68,16 +71,20 @@ public abstract class AbstractSpscBlockingQueueV2<E> extends AbstractBlockingQue
   public E poll() {
     final int headSeq = this.head.get();
     E e = null;
-    // 检查队列中是否有元素.
+    // 检查队列中是否有元素.   检查是否被其他线程修改,如果返回true,则没有修改,可以更新值.
     if (0 > headSeq || this.tail.get() > headSeq) {
-      // 获取环形数组索引位置.
-      final int bufferIndex = headSeq & index;
-      // 根据索引位置获取元素.
-      e = this.ringBuffer[bufferIndex];
-      // 将索引位置设置为空.
-      this.ringBuffer[bufferIndex] = null;
-      // poll计数.
-      this.head.getAndIncrement();
+      if (this.headLock.compareAndSet(headSeq, headSeq + 1)) {
+        // 获取环形数组索引位置.
+        final int bufferIndex = headSeq & index;
+        // 根据索引位置获取元素.
+        e = this.ringBuffer[bufferIndex];
+        // 将索引位置设置为空.
+        this.ringBuffer[bufferIndex] = null;
+        // poll计数.
+        this.head.getAndIncrement();
+      } else {
+        LockSupport.parkNanos(1L);
+      }
     }
     return e;
   }
@@ -104,7 +111,7 @@ public abstract class AbstractSpscBlockingQueueV2<E> extends AbstractBlockingQue
    * @return E e.
    */
   @Override
-  public final E peek() {
+  public E peek() {
     return ringBuffer[head.get() & index];
   }
 
@@ -117,7 +124,7 @@ public abstract class AbstractSpscBlockingQueueV2<E> extends AbstractBlockingQue
    * @return int e.
    */
   @Override
-  public final int size() {
+  public int size() {
     return Math.max(tail.get() - head.get(), 0);
   }
 
@@ -130,7 +137,7 @@ public abstract class AbstractSpscBlockingQueueV2<E> extends AbstractBlockingQue
    * @return boolean e.
    */
   @Override
-  public final boolean isEmpty() {
+  public boolean isEmpty() {
     return tail.get() == head.get();
   }
 }
