+++ jdkringbuffer-core/src/main/java/org/jdkstack/jdkringbuffer/core/spmc/version1/AbstractSpmcBlockingQueueV1.java
@@ -1,12 +1,6 @@
 package org.jdkstack.jdkringbuffer.core.spmc.version1;
 
-import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.locks.LockSupport;
-import org.jdkstack.jdkringbuffer.api.RingBufferBlockingQueue;
-import org.jdkstack.jdkringbuffer.core.AbstractBlockingQueue;
-import org.jdkstack.jdkringbuffer.core.Constants;
+import org.jdkstack.jdkringbuffer.core.AbstractLockBlockingQueueV1;
 
 /**
  * 单生产多消费SPMC阻塞队列.
@@ -16,22 +10,10 @@ import org.jdkstack.jdkringbuffer.core.Constants;
  * @author admin
  * @param <E> e .
  */
-public abstract class AbstractSpmcBlockingQueueV1<E> extends AbstractBlockingQueue<E>
-    implements BlockingQueue<E>, RingBufferBlockingQueue {
-  /** 环形数组. */
-  private final E[] ringBuffer;
-  /** 环形数组出队时,是否被其他线程抢先获取了值. */
-  private final AtomicInteger headLock = new AtomicInteger(0);
+public abstract class AbstractSpmcBlockingQueueV1<E> extends AbstractLockBlockingQueueV1<E> {
 
-  protected AbstractSpmcBlockingQueueV1() {
-    this(Constants.CAPACITY);
-  }
-
-  @SuppressWarnings("unchecked")
   protected AbstractSpmcBlockingQueueV1(final int capacity) {
-    super(capacity, capacity - 1);
-    // jdk 泛型数组,会有检查异常,但不影响什么,用unchecked关闭检查.
-    this.ringBuffer = (E[]) new Object[capacity];
+    super(capacity);
   }
 
   /**
@@ -58,87 +40,5 @@ public abstract class AbstractSpmcBlockingQueueV1<E> extends AbstractBlockingQue
       flag = true;
     }
     return flag;
-  }
-
-  /**
-   * This is a method description.
-   *
-   * <p>Another description after blank line.
-   *
-   * @author admin
-   * @return E e.
-   */
-  @Override
-  public E poll() {
-    final int headSeq = this.head.get();
-    E e = null;
-    // 检查队列中是否有元素.   检查是否被其他线程修改,如果返回true,则没有修改,可以更新值.
-    if (0 > headSeq || this.tail.get() > headSeq) {
-      if (this.headLock.compareAndSet(headSeq, headSeq + 1)) {
-        // 获取环形数组索引位置.
-        final int bufferIndex = headSeq & index;
-        // 根据索引位置获取元素.
-        e = this.ringBuffer[bufferIndex];
-        // 将索引位置设置为空.
-        this.ringBuffer[bufferIndex] = null;
-        // poll计数.
-        this.head.getAndIncrement();
-      } else {
-        LockSupport.parkNanos(1L);
-      }
-    }
-    return e;
-  }
-
-  /**
-   * This is a class description.
-   *
-   * <p>Another description after blank line.
-   *
-   * @author admin
-   * @return E e.
-   */
-  @Override
-  public E poll(final long timeout, final TimeUnit unit) {
-    throw new UnsupportedOperationException("未实现.");
-  }
-
-  /**
-   * This is a class description.
-   *
-   * <p>Another description after blank line.
-   *
-   * @author admin
-   * @return E e.
-   */
-  @Override
-  public final E peek() {
-    return ringBuffer[head.get() & index];
-  }
-
-  /**
-   * This is a class description.
-   *
-   * <p>Another description after blank line.
-   *
-   * @author admin
-   * @return int e.
-   */
-  @Override
-  public final int size() {
-    return Math.max(tail.get() - head.get(), 0);
-  }
-
-  /**
-   * This is a class description.
-   *
-   * <p>Another description after blank line.
-   *
-   * @author admin
-   * @return boolean e.
-   */
-  @Override
-  public final boolean isEmpty() {
-    return tail.get() == head.get();
   }
 }
